requirements.txt
streamlit==1.31.0
langchain==0.1.0
langchain-google-genai==0.0.8
langchain-community==0.0.20
weaviate-client==3.26.7
PyMuPDF==1.26.3
firebase-admin==6.4.0
arxiv==2.1.0
python-dotenv==1.0.1
streamlit-authenticator==0.3.1
pandas==2.1.4
numpy==1.26.4
tiktoken==0.6.0
requests==2.31.0
langchain-core==0.1.0
langchain-text-splitters==0.0.1
